fix: 경고 수준 4 적용, 경고를 오류로 처리, 전체 컴파일 경고 정리 - 미사용 변수, 매개변수, 함수 경고 해결 - 이름 숨김 및 중복 선언 경고 해결 - 정수 및 실수 형 변환 경고 수정 - 패킷 크기 범위 검증 및 예외 처리 추가 - 가변 인자 타입 전달 오류 수정 - 프로젝트 중간 디렉터리 충돌 및 기타 빌드 경고 해결 - 기타 C4002, C4005, C4091, C4100, C4101, C4102, C4189,   C4244, C4267, C4305, C4390, C4456, C4458, C4477,     C4505, C4840 및 MSB8028 경고 해결

+++ LL2_Client_Win_Source/stbCircleCollider2D.cpp
@@ -53,11 +53,12 @@ namespace stb
 		rightBottom.x = pos.x + offset.x + 100 * GetSize().x;
 		rightBottom.y = pos.y + offset.y + 100 * GetSize().y;
 
-		Ellipse(hdc
-			, pos.x + offset.x
-			, pos.y + offset.y
-			, rightBottom.x
-			, rightBottom.y);
+		Ellipse(
+			hdc,
+			static_cast<int>(pos.x + offset.x),
+			static_cast<int>(pos.y + offset.y),
+			static_cast<int>(rightBottom.x),
+			static_cast<int>(rightBottom.y));
 
 		SelectObject(hdc,oldBrush);
 		SelectObject(hdc, oldPen);
